Nanle - managing a growing project
